annotation_tool/requirements.txt
pandas==1.1.5
numpy==1.19.5
streamlit==1.9.0
st-annotated-text==3.0.0
protobuf~=3.19.0
